Split task classes into their own cpp/h files to speed up incremental compilation and make things easier to maintain
